+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree Android Drop-In Release Notes
 
+## unreleased
+
+* Bump braintree_android module dependency versions to `4.15.0`
+* Notify merchant via error message the payment method is duplicated (fixes #357)
+
 ## 6.3.0
 
 * Deprecate constructors that require a `DropInRequest`.